+++ Changelog.md
@@ -13,6 +13,7 @@
 * [Use `unsafeWithForeignPtr` whenever possible](https://github.com/haskell/bytestring/pull/401)
 * [Remove `integer-simple` flag](https://github.com/haskell/bytestring/pull/371)
 * [Remove misleading mentions of fusion](https://github.com/haskell/bytestring/pull/412)
+* Add `Data.ByteString.isValidUtf8`
 
 [0.11.2.0]: https://github.com/haskell/bytestring/compare/0.11.1.0...0.11.2.0